Add files via upload
This is my one of my college projects written in python. The data.txt and invertex_index files are shortened versions of the real files. The full files were too big to upload to github (around 200mb each). In the code the part that is commented out is where the program creates the inverted index given data.txt and saves it as a file -> inverted_index1.txt. This normally would take hours with the actual data.txt file and also a reason why in the program I chose to save the inverted index after creation so that you wouldn't have to wait hours to start searching words on runtime. The program would just load inverted_index1.txt and you could start searching.